refactored resource cards
