+++ app/_components/(workshop)/additional-resources/cards.json
@@ -26,7 +26,7 @@
     {
         "title": "HTML Semantics",
         "type": "Website", 
-        "image": "/images/filler.jpeg",
+        "image": "",
         "link": "https://flukeout.github.io/"
     },
     {
