--- app/_components/(workshop)/workshop-additional-resources/additional-resources.module.scss
@@ -1,37 +1,0 @@
-.mainBodyContainer {
-    background-color: var(--background-light);
-    padding-bottom: 325px;
-    
-    h3 {
-        font-size: 3.25rem;
-        color: var( --text-midtone);
-        padding-top: 114px;
-        padding-left: 145px; 
-        padding-bottom: 85px;
-    }
-}
-
-.tab{
-    display: flex;
-    width: 374px; 
-    height: 100px; 
-    background-color: var(--background-light);
-    clip-path: polygon(0 0, calc(100% - 74px) 0, 100% 100px, 100% 100%, 0 100%);
-}
-
-
-.cardsContainer {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 30px; 
-    justify-content: center;
-}
-
-.videoContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 50px;
-    padding-left: 145px;
-    padding-bottom: 104px;
-    
-}
