--- app/_components/(workshop)/workshop-additional-resources/card.module.scss
@@ -1,35 +1,0 @@
-.card {
-    display: flex;
-    flex-direction: column;
-    gap: 35px; 
-    background-color: var( --include-purple);
-    width: 463px; 
-    height: 325px; 
-    border-radius: 20px;
-    padding: 35px 35px 30px 35px; 
-    color: var(--text-light);
-    h2 {
-        font-size: 2rem;
-    }
-
-    p {
-        font-size: 1.5rem;
-        font-weight: 400;
-    }
-}
-
-.textContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 7px;
-}
-
-.imgContainer {
-    position: relative; 
-    width: 395px; 
-    height: 150px; 
-    border-radius: 20px;
-    overflow: hidden;
-    opacity: 80%;
-
-}
