--- app/_components/(workshop)/workshop-additional-resources/video-player.module.scss
@@ -1,25 +1,0 @@
-.videoContainer {
-    display: flex;
-    gap: 58px;
-}
-
-.textContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 10px; 
-    width: 625px; 
-    
-    h3 {
-        
-        font-size: 1.5rem;
-        padding: 0; 
-        color: var(--text-midtone);
-        font-weight: 700;
-        
-    }
-    p {
-        font-size: 1.5rem;
-        color: var(--text-midtone);
-        font-weight: 300;
-    }
-}
