--- app/_components/(workshop)/workshop-additional-resources/videos.json
@@ -1,13 +1,0 @@
-[
-    {
-        "title": "HTML Basics",
-        "description": "Watch the video above for more in-depth information regarding coding in HTML.",
-        "videoSrc": "https://www.youtube.com/embed/0W6qz0-aDaM?si=Nfo5iC-dYyLwUaox"
-    }, 
-    {
-        "title": "Introduction to CSS",
-        "description": "Watch the video above for more in-depth information writing CSS.",
-        "videoSrc": "https://www.youtube.com/embed/0W6qz0-aDaM?si=Nfo5iC-dYyLwUaox"
-
-    }
-]
