+++ next.config.mjs
@@ -4,7 +4,7 @@ const nextConfig = {
         remotePatterns: [
           {
             protocol: "https",
-            hostname: "*.googleusercontent.com",
+            hostname: "*",
           },
         ],
     },
